feat: 레이아웃 통합
/plan/schedule, plan/store/plan, plan/store/place
