+++ app/(route)/plan/(addPlace)/schedule/page.tsx
@@ -3,53 +3,14 @@ import React, { ReactNode, useState } from 'react'
 import { Map } from 'react-kakao-maps-sdk'
 
 import { SpriteMapMarker } from '@/components/common/MapPin'
-import AddedPlaceLists from '@/components/plan/AddedPlaceLists'
 import SearchArea from '@/components/plan/SerachArea'
 import { Button } from '@/components/ui/button'
 import { BACKEND_ROUTES } from '@/lib/constants/routes'
 import usePlanStore from '@/lib/context/planStore'
 import { Place } from '@/lib/types/Entity/place'
-import { DayPlan } from '@/lib/types/Entity/plan'
 import useKakaoLoader from '@/lib/utils/hooks/useKakaoLoader'
-import DUMMYPLACEIMG from '@/public/dummy/dummy_place_image.png'
 
 interface PlanSchedulePageProps {}
-
-// Dummy Place
-const DUMMY_PLACE: Place = {
-  id: 125405,
-  name: '토함산자연휴양림',
-  imgSrc: DUMMYPLACEIMG,
-  address: '경상북도 경주시 양북면 불국로',
-
-  tag: '관광지',
-  duration: 60,
-  stars: 5,
-  visitCnt: 100, // 실제 계획에 담긴 횟수
-
-  geo: {
-    latitude: 33.450701,
-    longitude: 126.570667,
-  },
-  //reviews: Array<Comment>
-  reviewCnt: 10,
-
-  isAdded: true, // 계획에 들어갔는지 여부
-  isScraped: true,
-
-  order: 1, // 계획세우기에 담긴 순서
-}
-
-// Dummy places
-const DUMMY_PLACES: Array<Place> = new Array(12).fill(DUMMY_PLACE)
-
-// Dummy DayPlan
-export const DUMMY_DAYPLAN: DayPlan = {
-  day: 1,
-  startTime: '08:00',
-  endTime: '20:00',
-  places: DUMMY_PLACES,
-}
 
 const PlanSchedulePage = ({}: PlanSchedulePageProps): ReactNode => {
   useKakaoLoader() // 카카오 지도 로딩
@@ -85,12 +46,10 @@ const PlanSchedulePage = ({}: PlanSchedulePageProps): ReactNode => {
     // update();
     setFocusedPlaceCard(undefined) // 초기화
   }
-  // Todo: DayPlan 정보 받아오기
-  const DayPlan: DayPlan = DUMMY_DAYPLAN
 
   return (
-    <div className='relative flex h-dvh flex-grow justify-start'>
-      <AddedPlaceLists DayPlan={DayPlan} />
+    <>
+      {/* <AddedPlaceLists DayPlan={DayPlan} /> */}
 
       {/* 검색창 */}
       {isSearching && (
@@ -144,7 +103,7 @@ const PlanSchedulePage = ({}: PlanSchedulePageProps): ReactNode => {
           </Button>
         )}
       </div>
-    </div>
+    </>
   )
 }
