+++ components/plan/AddedPlaceLists.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { ReactNode } from 'react'
 
 import usePlanStore from '@/lib/context/planStore'
